requirements.txt
pillow==11.0.0
fastapi==0.115.5
pydantic==2.10.0
uvicorn==0.32.1
sentencepiece==0.2.0
tokenizers==0.20.3
opencv-python==4.10.0.84
python-multipart==0.0.17
fastapi-cors==0.0.6
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
gunicorn==23.0.0
onnxruntime-gpu==1.20.0
transformers==4.46.3

